pytest extract tests
