+++ requirements.txt
@@ -1,8 +1,13 @@
 certifi==2025.4.26
 charset-normalizer==3.4.2
 idna==3.10
+iniconfig==2.1.0
+packaging==25.0
+pluggy==1.6.0
 py4j==0.10.9.9
 pyspark==4.0.0
+pytest==8.3.5
+pytest-mock==3.14.1
 python-dotenv==1.1.0
 PyYAML==6.0.2
 requests==2.32.3
